+++ angel_serialize_generator/test/models/game_pad_button.dart
@@ -1,7 +1,7 @@
 import 'package:angel_serialize/angel_serialize.dart';
 part 'game_pad_button.g.dart';
 
-@Serializable(autoIdAndDateFields: false)
+@serializable
 abstract class _GamepadButton {
   String get name;
   int get radius;