+++ angel_serialize_generator/test/models/goat.dart
@@ -2,7 +2,7 @@ import 'package:angel_serialize/angel_serialize.dart';
 import 'package:collection/collection.dart';
 part 'goat.g.dart';
 
-@Serializable(autoIdAndDateFields: false)
+@serializable
 abstract class _Goat {
   @SerializableField(defaultValue: 34)
   int get integer;